Separate out Parse method for cleaner code.
