+++ libs/server/Resp/Objects/SortedSetGeoCommands.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 
 using System;
-using System.Text;
 using Garnet.common;
 using Tsavorite.core;
 
@@ -160,7 +159,7 @@ namespace Garnet.server
             var paramsRequiredInCommand = 0;
             var searchOpts = new GeoSearchOptions();
 
-            int sourceIdx = 0, destIdx = -1;
+            int sourceIdx = 0;
             bool readOnly = false;
             switch (command)
             {
@@ -184,7 +183,6 @@ namespace Garnet.server
                     break;
                 case RespCommand.GEOSEARCHSTORE:
                     paramsRequiredInCommand = 7;
-                    destIdx = 0;
                     sourceIdx = 1;
                     break;
                 default:
@@ -206,363 +204,13 @@ namespace Garnet.server
             }, ref parseState, startIdx: sourceIdx + 1);
             var output = new SpanByteAndMemory(dcurr, (int)(dend - dcurr));
 
-            #region parse
-            ReadOnlySpan<byte> errorMessage = default;
-            var argNumError = false;
-            var storeDist = false;
-            var currTokenIdx = 0;
-
-            if (command == RespCommand.GEORADIUS || command == RespCommand.GEORADIUS_RO ||
-                command == RespCommand.GEORADIUSBYMEMBER || command == RespCommand.GEORADIUSBYMEMBER_RO)
-            {
-                // Read coordinates, note we already checked the number of arguments earlier.
-                if (command == RespCommand.GEORADIUSBYMEMBER || command == RespCommand.GEORADIUSBYMEMBER_RO)
-                {
-                    // From Member
-                    searchOpts.fromMember = input.parseState.GetArgSliceByRef(currTokenIdx++).SpanByte.ToByteArray();
-                    searchOpts.origin = GeoOriginType.FromMember;
-                }
-                else
-                {
-                    if (!input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.lon) ||
-                        !input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.lat))
-                    {
-                        while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_NOT_VALID_FLOAT, ref dcurr, dend))
-                            SendAndReset();
-                        return true;
-                    }
-
-                    if ((searchOpts.lon < GeoHash.LongitudeMin) ||
-                        (searchOpts.lat < GeoHash.LatitudeMin) ||
-                        (searchOpts.lon > GeoHash.LongitudeMax) ||
-                        (searchOpts.lat > GeoHash.LatitudeMax))
-                    {
-                        errorMessage = Encoding.ASCII.GetBytes(string.Format(CmdStrings.GenericErrLonLat,
-                                                               searchOpts.lon, searchOpts.lat));
-                        while (!RespWriteUtils.TryWriteError(errorMessage, ref dcurr, dend))
-                            SendAndReset();
-                        return true;
-                    }
-
-                    searchOpts.origin = GeoOriginType.FromLonLat;
-                }
-
-                // Radius
-                if (!input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.radius))
-                {
-                    while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_NOT_VALID_RADIUS, ref dcurr, dend))
-                        SendAndReset();
-                    return true;
-                }
-
-                if (searchOpts.radius < 0)
-                {
-                    while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_GENERIC_VALUE_IS_OUT_OF_RANGE, ref dcurr, dend))
-                        SendAndReset();
-                    return true;
-                }
-
-                searchOpts.searchType = GeoSearchType.ByRadius;
-                if (!input.parseState.TryGetGeoDistanceUnit(currTokenIdx++, out searchOpts.unit))
-                {
-                    while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_NOT_VALID_GEO_DISTANCE_UNIT, ref dcurr, dend))
-                        SendAndReset();
-                    return true;
-                }
-            }
-
-            // Read the options
-            while (currTokenIdx < input.parseState.Count)
-            {
-                // Read token
-                var tokenBytes = input.parseState.GetArgSliceByRef(currTokenIdx++).ReadOnlySpan;
-
-                if (command == RespCommand.GEOSEARCH || command == RespCommand.GEOSEARCHSTORE)
-                {
-                    if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.FROMMEMBER))
-                    {
-                        if (searchOpts.origin != GeoOriginType.Undefined)
-                        {
-                            errorMessage = CmdStrings.RESP_SYNTAX_ERROR;
-                            break;
-                        }
-
-                        if (input.parseState.Count == currTokenIdx)
-                        {
-                            argNumError = true;
-                            break;
-                        }
-
-                        searchOpts.fromMember = input.parseState.GetArgSliceByRef(currTokenIdx++).SpanByte.ToByteArray();
-                        searchOpts.origin = GeoOriginType.FromMember;
-                        continue;
-                    }
-
-                    if (tokenBytes.EqualsUpperCaseSpanIgnoringCase("FROMLONLAT"u8))
-                    {
-                        if (searchOpts.origin != GeoOriginType.Undefined)
-                        {
-                            errorMessage = CmdStrings.RESP_SYNTAX_ERROR;
-                            break;
-                        }
-
-                        if (input.parseState.Count - currTokenIdx < 2)
-                        {
-                            argNumError = true;
-                            break;
-                        }
-
-                        // Read coordinates
-                        if (!input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.lon) ||
-                            !input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.lat)
-                           )
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_NOT_VALID_FLOAT;
-                            break;
-                        }
-
-                        if ((searchOpts.lon < GeoHash.LongitudeMin) ||
-                            (searchOpts.lat < GeoHash.LatitudeMin) ||
-                            (searchOpts.lon > GeoHash.LongitudeMax) ||
-                            (searchOpts.lat > GeoHash.LatitudeMax))
-                        {
-                            errorMessage = Encoding.ASCII.GetBytes(string.Format(CmdStrings.GenericErrLonLat,
-                                                                   searchOpts.lon, searchOpts.lat));
-                            while (!RespWriteUtils.TryWriteError(errorMessage, ref dcurr, dend))
-                                SendAndReset();
-                            return true;
-                        }
-
-                        searchOpts.origin = GeoOriginType.FromLonLat;
-                        continue;
-                    }
-
-                    if (tokenBytes.EqualsUpperCaseSpanIgnoringCase("BYRADIUS"u8))
-                    {
-                        if (searchOpts.searchType != GeoSearchType.Undefined)
-                        {
-                            errorMessage = CmdStrings.RESP_SYNTAX_ERROR;
-                            break;
-                        }
-
-                        if (input.parseState.Count - currTokenIdx < 2)
-                        {
-                            argNumError = true;
-                            break;
-                        }
-
-                        // Read radius and units
-                        if (!input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.radius))
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_NOT_VALID_RADIUS;
-                            break;
-                        }
-
-                        if (searchOpts.radius < 0)
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_GENERIC_VALUE_IS_OUT_OF_RANGE;
-                            break;
-                        }
-
-                        searchOpts.searchType = GeoSearchType.ByRadius;
-                        if (!input.parseState.TryGetGeoDistanceUnit(currTokenIdx++, out searchOpts.unit))
-                        {
-                            while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_NOT_VALID_GEO_DISTANCE_UNIT, ref dcurr, dend))
-                                SendAndReset();
-                            return true;
-                        }
-                        continue;
-                    }
-
-                    if (tokenBytes.EqualsUpperCaseSpanIgnoringCase("BYBOX"u8))
-                    {
-                        if (searchOpts.searchType != GeoSearchType.Undefined)
-                        {
-                            errorMessage = CmdStrings.RESP_SYNTAX_ERROR;
-                            break;
-                        }
-                        searchOpts.searchType = GeoSearchType.ByBox;
-
-                        if (input.parseState.Count - currTokenIdx < 3)
-                        {
-                            argNumError = true;
-                            break;
-                        }
-
-                        // Read width, height
-                        if (!input.parseState.TryGetDouble(currTokenIdx++, out searchOpts.boxWidth))
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_NOT_VALID_WIDTH;
-                            break;
-                        }
-
-                        if (!input.parseState.TryGetDouble(currTokenIdx++, out var height))
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_NOT_VALID_HEIGHT;
-                            break;
-                        }
-
-                        searchOpts.boxHeight = height;
-
-                        if (searchOpts.boxWidth < 0 || searchOpts.boxHeight < 0)
-                        {
-                            errorMessage = CmdStrings.RESP_ERR_HEIGHT_OR_WIDTH_NEGATIVE;
-                            break;
-                        }
-
-                        // Read units
-                        if (!input.parseState.TryGetGeoDistanceUnit(currTokenIdx++, out searchOpts.unit))
-                        {
-                            while (!RespWriteUtils.TryWriteError(CmdStrings.RESP_ERR_NOT_VALID_GEO_DISTANCE_UNIT, ref dcurr, dend))
-                                SendAndReset();
-                            return true;
-                        }
-                        continue;
-                    }
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase("ASC"u8))
-                {
-                    searchOpts.sort = GeoOrder.Ascending;
-                    continue;
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase("DESC"u8))
-                {
-                    searchOpts.sort = GeoOrder.Descending;
-                    continue;
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.COUNT))
-                {
-                    if (input.parseState.Count == currTokenIdx)
-                    {
-                        argNumError = true;
-                        break;
-                    }
-
-                    if (!input.parseState.TryGetInt(currTokenIdx++, out var countValue))
-                    {
-                        errorMessage = CmdStrings.RESP_ERR_GENERIC_VALUE_IS_NOT_INTEGER;
-                        break;
-                    }
-
-                    if (countValue <= 0)
-                    {
-                        errorMessage = CmdStrings.RESP_ERR_COUNT_IS_NOT_POSITIVE;
-                        break;
-                    }
-
-                    searchOpts.countValue = countValue;
-
-                    if (input.parseState.Count > currTokenIdx)
-                    {
-                        var peekArg = input.parseState.GetArgSliceByRef(currTokenIdx).ReadOnlySpan;
-                        if (peekArg.EqualsUpperCaseSpanIgnoringCase("ANY"u8))
-                        {
-                            searchOpts.withCountAny = true;
-                            currTokenIdx++;
-                            continue;
-                        }
-                    }
-
-                    continue;
-                }
-
-                if (!readOnly)
-                {
-                    if ((command == RespCommand.GEORADIUS || command == RespCommand.GEORADIUSBYMEMBER)
-                        && tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.STORE))
-                    {
-                        if (input.parseState.Count == currTokenIdx)
-                        {
-                            argNumError = true;
-                            break;
-                        }
-
-                        destIdx = ++currTokenIdx;
-                        continue;
-                    }
-
-                    if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.STOREDIST))
-                    {
-                        if ((command == RespCommand.GEORADIUS || command == RespCommand.GEORADIUSBYMEMBER))
-                        {
-                            if (input.parseState.Count == currTokenIdx)
-                            {
-                                argNumError = true;
-                                break;
-                            }
-
-                            destIdx = ++currTokenIdx;
-                        }
-
-                        storeDist = true;
-                        continue;
-                    }
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.WITHCOORD))
-                {
-                    searchOpts.withCoord = true;
-                    continue;
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.WITHDIST))
-                {
-                    searchOpts.withDist = true;
-                    continue;
-                }
-
-                if (tokenBytes.EqualsUpperCaseSpanIgnoringCase(CmdStrings.WITHHASH))
-                {
-                    searchOpts.withHash = true;
-                    continue;
-                }
-
-                errorMessage = CmdStrings.RESP_SYNTAX_ERROR;
-                break;
-            }
-
-            // Check that we have the mandatory options
-            if (errorMessage.IsEmpty && ((searchOpts.origin == 0) || (searchOpts.searchType == 0)))
-                argNumError = true;
-
-            // Check if we have a wrong number of arguments
-            if (argNumError)
-            {
-                errorMessage = Encoding.ASCII.GetBytes(string.Format(CmdStrings.GenericErrWrongNumArgs,
-                                                       command.ToString()));
-            }
-
-            if (destIdx != -1)
-            {
-                if ((searchOpts.withDist || searchOpts.withCoord || searchOpts.withHash))
-                {
-                    errorMessage = Encoding.ASCII.GetBytes(string.Format(CmdStrings.GenericErrStoreCommand,
-                                                           command.ToString()));
-                }
-                searchOpts.withDist = storeDist;
-            }
-            else
-            {
-                readOnly = true;
-            }
-
-            // Check if we encountered an error while checking the parse state
+            var errorMessage = GeoSearchOptions.Parse(command, ref searchOpts, ref input, readOnly, out var destIdx);
             if (!errorMessage.IsEmpty)
             {
                 while (!RespWriteUtils.TryWriteError(errorMessage, ref dcurr, dend))
                     SendAndReset();
                 return true;
             }
-
-            // On storing to ZSET, we need to use either dist or hash as score.
-            if (!readOnly && !searchOpts.withDist && !searchOpts.withHash)
-            {
-                searchOpts.withHash = true;
-            }
-            #endregion
 
             GarnetStatus status;
             if (destIdx != -1)
